Implementation JWT - Refonte BDD

+++ Sama_Travel/src/main/java/fr/dawan/SamaTravel/repositories/AppUserRepository.java
@@ -1,13 +1,10 @@
 package fr.dawan.SamaTravel.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
 import fr.dawan.SamaTravel.entities.AppUser;
 
-public interface AppUserRepository extends JpaRepository<AppUser, Integer> {
-	
-	
+public interface AppUserRepository extends JpaRepository<AppUser, Long> {
 	
 	public AppUser findByUsername(String username);
